Add float to get values between 0 and 200

+++ player/Python/modules/_classes.py
@@ -428,7 +428,7 @@ class AbstractVLC(ExternalProcessFlag):
          :return: Absolute volume for VLC between 0 and 1024
          :rtype: int
         """
-        return int(int(settings.get("vlc", "volume", "master")) * (int(volume)/100))
+        return int(float(settings.get("vlc", "volume", "master")) * (float(volume)/100))
 
     def set_volume(self, volume):
         """
